Fixing type return pourcent

+++ math/classic.go
@@ -1,8 +1,8 @@
 package math
 
 // Pourcent value
-func Pourcent(a, b int64) float32 {
-	return (float32(a) / float32(b)) / 100
+func Pourcent(a, b int64) float64 {
+	return (float64(a) / float64(b)) / 100
 }
 
 // Diff value
